refactored exporter to match the clinverter and not initialize its own mysql client
