+++ Gemfile
@@ -2,7 +2,12 @@ source 'https://rubygems.org'
 
 gem 'rake'
 gem 'zeitwerk'
+
+# exporter
 gem 'mysql2'
+
+# converter
+gem 'aws-sdk-s3'
 
 group :development, :test do
   gem 'minitest'
