+++ bin/exporter
@@ -1,18 +1,18 @@
 #!/usr/bin/env ruby
 
 require 'zeitwerk'
+require 'mysql2'
 require_relative "../lib/aurora_bootstrapper"
 
-db_user   = ENV.fetch "DB_USER"
-db_pass   = ENV.fetch "DB_PASS"
-db_host   = ENV.fetch "DB_HOST"
 prefix    = ENV.fetch "PREFIX", ""
 bukkit    = ENV.fetch "EXPORT_BUCKET"
 blacklist = ENV.fetch "BLACKLISTED_TABLES", ""
 
-AuroraBootstrapper::Exporter.new( db_host: db_host,
-                                  db_user: db_user,
-                                  db_pass: db_pass,
-                                   prefix: prefix,
-                            export_bucket: bukkit,
-                       blacklisted_tables: blacklist ).export!
+client    = Mysql2::Client.new( host: ENV.fetch( "DB_USER" ),
+                            username: ENV.fetch( "DB_PASS" ),
+                            password: ENV.fetch( "DB_HOST" ))
+
+AuroraBootstrapper::Exporter.new( client: client,
+                                  prefix: prefix,
+                           export_bucket: bukkit,
+                      blacklisted_tables: blacklist ).export!
