+++ lib/aurora_bootstrapper.rb
@@ -1,4 +1,3 @@
-require "mysql2"
 require "logger"
 require "json"
