+++ test/lib/converter_test.rb
