+++ test/test_helper.rb
@@ -1,6 +1,7 @@
 $LOAD_PATH.unshift File.expand_path("../../lib", __FILE__)
 require 'simplecov'
 require 'dotenv'
+require 'mysql2'
 
 Dotenv.load('.env.test')
